fix(fanout): a worker's bibliography is not lost for arriving late
`merge_worker_sources` built one block of every source a worker found that the
parent did not have, and handed it to `add_to_region`, which refuses an
over-budget entry whole. The refusal was discarded. So once the parent's
bibliography region neared its cap, every later worker's sources were dropped in
silence and the run reported success.

Measured on the finished 7-worker deep-researcher run: `sources_index` ended at
19,461 tokens of 20,000, holding the parent's own bibliography and **four**
worker merges. The sources found by the other three are nowhere in the run, and
nothing anywhere said so. The report cited 45 URLs; the workers found 376.

Two changes, both the same idea:

  * Contribute what fits. The block is trimmed to the region's remaining
    headroom, line by line, so a worker that cannot fit all its sources still
    contributes the ones there is room for rather than none.
  * Say what did not fit, naming the worker and the count. A dropped source is
    one the report cannot cite, which is the difference between a bibliography
    and a guess, and it names the budget to raise.

The write is handed the same running total the headroom check used, rather than a
fresh estimate of the joined block. Two numbers for one quantity is what made the
old code need an error arm to reconcile them; one number makes the write fit by
construction.

`fanout.rs` went over the 1200-line rule with this, so the subject moved to
`fanout/worker_sources.rs` with its tests. It is a separable one: that module is
how a fan-out is split, run and merged, this one is bibliographies and
deduplication by URL, and they meet at a single call when a worker is reaped.

The regression test was run against the unfixed code first and fails there with
"a full region used to take none of them".
